Base refactor complete. Next Post Search: FULL not Recent.
