+++ routes/twitter.js
@@ -8,9 +8,7 @@ const dotenv = require("dotenv");
 dotenv.config();
 const beefDap = process.env.BEEF_URI;
 
-
 const userSearchFields = "user.fields=created_at,name,id,profile_image_url,verified,verified_type";
-
 
 
 const performUserSearch = async (users, useSession = true) => {
@@ -196,7 +194,139 @@ const twitterRevokeHandler = async (req, res) => {
     res.send('OK');
 }
 
+const twitterRecentSearchHandler = async (req, res) => {
 
+    const { xt, search, start, end } = req.query;
+
+    if ((req.session.at || xt) && req.session[generateMD5Hash(search)] == null) {
+        //  v2 Auth Pattern         
+        const timeRange = (start ? '&' + start : '') + (end ? '&' + end : '');
+        const searchResponse = await axios.get("https://api.x.com/2/tweets/search/recent?query=" + search + timeRange + "&tweet.fields=created_at&expansions=author_id&max_results=100&" + userSearchFields, {
+            headers: {
+                "User-Agent": "v2RecentSearchJS",
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${xt}`,
+            },
+        }).catch((err) => {
+            res.send(JSON.stringify(handleXAPIErrors(err)));
+            return
+        });
+
+        if (searchResponse) {
+            req.session[generateMD5Hash(search)] = JSON.stringify(searchResponse.data);
+            res.send(JSON.stringify(searchResponse.data));
+        }
+        return;
+
+
+    } else if (req.session[generateMD5Hash(search)] != null) {
+        console.info("Using Session");
+        res.send(req.session[generateMD5Hash(search)]);
+    } else {
+        res.send(JSON.stringify({ error: 'X SEARCH ERROR: Error', login: 0 }));
+    }
+
+}
+
+const twitterSearchHandler = async (req, res) => {
+
+    const { xt, search, start, end } = req.query;
+
+    if ((req.session.at || xt) && req.session[generateMD5Hash(search)] == null) {
+
+
+        //  v2 Auth Pattern         
+        const timeRange = (start ? '&' + start : '') + (end ? '&' + end : '');
+        const searchResponse = await axios.get("https://api.x.com/2/tweets/search/all?query=" + search + timeRange + "&tweet.fields=created_at&expansions=author_id&max_results=100&" + userSearchFields, {
+            headers: {
+                "User-Agent": "v2RecentSearchJS",
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${process.env.X_BEARER_TOKEN}`,
+            },
+        }).catch((err) => {
+            res.send(JSON.stringify(handleXAPIErrors(err)));
+            return
+        });
+
+        if (searchResponse) {
+            req.session[generateMD5Hash(search)] = JSON.stringify(searchResponse.data);
+            res.send(JSON.stringify(searchResponse.data));
+        }
+        return;
+
+
+    } else if (req.session[generateMD5Hash(search)] != null) {
+        console.info("Using Session");
+        res.send(req.session[generateMD5Hash(search)]);
+    } else {
+        res.send(JSON.stringify({ error: 'X SEARCH ERROR: Error', login: 0 }));
+    }
+}
+
+const twitterUserSearchHandler = async (req, res) => {
+    const { users } = req.query;
+
+    if ((req.session.at) && req.session[generateMD5Hash(users)] == null) {
+
+        if (users && users.length > 0) {
+            try {
+                const searched = await performUserSearch(users)
+                req.session[generateMD5Hash(users)] = searched;
+                res.send(searched);
+            } catch (err) {
+                res.send(JSON.stringify({ error: 'X SEARCH ERROR: Login', login: 1 }));
+            }
+
+        } else {
+            res.send(JSON.stringify({ error: 'X SEARCH ERROR: NO Params', login: 0 }));
+        }
+
+    } else if (req.session[generateMD5Hash(users)] != null) {
+        console.info("loading from cache");
+        // saved as JSON STRING
+        res.send(req.session[generateMD5Hash(users)]);
+    } else {
+        if (users && users.length > 0) {
+            // User Search Application Search
+            const searched = await performUserSearch(users, false);
+            req.session[generateMD5Hash(users)] = searched;
+            res.send(searched);
+        } else {
+            res.send(JSON.stringify({ error: 'X SEARCH ERROR: NO Params', login: 0 }));
+        }
+    }
+}
+
+const twitterPostSearchHandler = async (req, res) => {
+
+    const { xt, id } = req.query;
+    // Keep for Session saver for dev purposes
+    // req.session[generateMD5Hash(id)] = { "data": { "id": "1910392237394972890", "edit_history_tweet_ids": ["1910392237394972890"], "author_id": "1393563533820977159", "text": "CASTLES ARE BETTER", "created_at": "2025-04-10T17:59:37.000Z" }, "includes": { "users": [{ "profile_image_url": "https://pbs.twimg.com/profile_images/1415399629622059012/7J2sLEPz_normal.jpg", "verified": false, "verified_type": "none", "name": "WOBInteractive", "created_at": "2021-05-15T13:47:17.000Z", "id": "1393563533820977159", "username": "webofblood1" }] } };
+
+    if ((req.session.at || xt) && id && req.session[generateMD5Hash(id)] == null) {
+        const searchResponse = await axios.get("https://api.x.com/2/tweets/" + id + "?tweet.fields=created_at,text&expansions=author_id&" + userSearchFields, {
+            headers: {
+                "User-Agent": "v2RecentSearchJS",
+                "Content-Type": "application/json",
+                Authorization: `Bearer ${xt}`,
+            },
+        }).catch((err) => {
+            res.send(JSON.stringify(handleXAPIErrors(err)));
+        });
+
+        if (searchResponse) {
+            req.session[generateMD5Hash(id)] = searchResponse.data;
+            req.session.t = xt;
+            res.send(JSON.stringify(searchResponse.data));
+        }
+    } else if (id && req.session[generateMD5Hash(id)] != null) {
+        console.info("Using Session");
+        res.send(req.session[generateMD5Hash(id)]);
+    } else {
+        res.send(JSON.stringify({ error: 'X SEARCH ERROR', login: xt ? 0 : 1, code: 401 })); // unauth'd
+    }
+    return
+}
 
 
 module.exports = {
@@ -205,5 +335,9 @@ module.exports = {
     handleXAPIErrors,
     twitterLogInHandler,
     twitterLoginCallbackHandler,
-    twitterRevokeHandler
+    twitterRevokeHandler,
+    twitterRecentSearchHandler,
+    twitterSearchHandler,
+    twitterUserSearchHandler,
+    twitterPostSearchHandler
 };
