--- routes/user.js
@@ -1,8 +1,0 @@
-// routes/user.js
-const { getXUserData, handleXAPIErrors } = require('../routes/auth');
-const { performUserSearch } = require('./twitter');
-
-module.exports = {
-    getXUserData,
-    handleXAPIErrors,
-};
